Package structure refactored
